server/requirements.txt
Django
psycopg2
psycopg2-binary
django-oauth-toolkit
djangorestframework
django-cors-headers
django-rest-swagger
django-filter
rest_condition
boto3
uwsgi
drf-yasg
packaging
django-constance[database]
flake8
coverage
coreapi
pyyaml
python-dotenv
